More refactoring work
